rename .cc files to .cpp and run uncrustify
ament_uncrustify is configured to treat .cc files as plain C, and .cpp
files as C++.

this commit renames all .cc files to .cpp and applies a fresh coat of
uncrustify to them.
the layout is much nicer now that uncrustify knows
there's namespaces and templates involved!
