+++ icon_hwm_controller/test/domain_socket_utils_test.cpp
@@ -14,22 +14,24 @@
 #include "intrinsic/utils/status.hpp"
 #include "intrinsic/utils/time.hpp"
 
-namespace intrinsic::hal {
+namespace intrinsic::hal
+{
 
-  namespace {
+namespace
+{
 
-    using::testing::FieldsAre;
-    using::testing::HasSubstr;
-    using::testing::TempDir;
+using ::testing::FieldsAre;
+using ::testing::HasSubstr;
+using ::testing::TempDir;
 
 // domain_socket_internal
-    TEST(CreateSocketDirectory, ErrorsOnRelativePath) {
+TEST(CreateSocketDirectory, ErrorsOnRelativePath) {
   EXPECT_THAT(domain_socket_internal::CreateSocketDirectory("tmp/some_dir"),
               FieldsAre(StatusCode::kInvalidArgument, HasSubstr("absolute")));
-    }
+}
 
 // domain_socket_internal
-    TEST(CreateSocketDirectory, Works) {
+TEST(CreateSocketDirectory, Works) {
       auto socket_dir = std::filesystem::path(TempDir()) / "some_dir";
 
   EXPECT_EQ(domain_socket_internal::CreateSocketDirectory(socket_dir).code, StatusCode::kOk);
@@ -37,28 +39,28 @@ namespace intrinsic::hal {
   std::error_code ec;
   EXPECT_TRUE(std::filesystem::exists(socket_dir, ec)) << ec;
   EXPECT_TRUE(std::filesystem::is_directory(socket_dir, ec)) << ec;
-    }
+}
 
   // domain_socket_internal
-    TEST(AbsoluteSocketPath, ErrorsOnRelativePath) {
+TEST(AbsoluteSocketPath, ErrorsOnRelativePath) {
     auto result = domain_socket_internal::AbsoluteSocketPath("tmp/some_path",
                                                              "some_module");
     ASSERT_FALSE(result.has_value());
     EXPECT_THAT(
         result.error(),
         FieldsAre(StatusCode::kInvalidArgument, HasSubstr("absolute")));
-    }
+}
 
 // domain_socket_internal
-    TEST(AbsoluteSocketPath, Works) {
+TEST(AbsoluteSocketPath, Works) {
   auto result = domain_socket_internal::AbsoluteSocketPath("/tmp/some_dir",
                                                            "some_module");
   ASSERT_TRUE(result.has_value()) << result.error();
   EXPECT_EQ(result.value(), "/tmp/some_dir/some_module.sock");
-    }
+}
 
 // domain_socket_internal
-    TEST(AddressFromAbsolutePath, ChecksLength) {
+TEST(AddressFromAbsolutePath, ChecksLength) {
   std::string too_long_path(sizeof(sockaddr_un::sun_path) + 1, 'a');
 
   auto result = domain_socket_internal::AddressFromAbsolutePath(too_long_path);
@@ -66,12 +68,12 @@ namespace intrinsic::hal {
   EXPECT_THAT(
       result.error(),
       FieldsAre(StatusCode::kInvalidArgument, HasSubstr("too long")));
-    }
+}
 
 // domain_socket_internal
-    TEST(AddressFromAbsolutePath, Works) {
+TEST(AddressFromAbsolutePath, Works) {
   std::filesystem::path socket_path = (std::filesystem::path(TempDir()) / "some_socket")
-        .concat(domain_socket_internal::kSocketSuffix);
+    .concat(domain_socket_internal::kSocketSuffix);
 
   auto addr = domain_socket_internal::AddressFromAbsolutePath(socket_path);
   ASSERT_TRUE(addr.has_value()) << addr.error();
@@ -80,9 +82,9 @@ namespace intrinsic::hal {
   EXPECT_EQ(addr->sun_path, socket_path);
   // Checks that the null terminator is set.
   EXPECT_EQ(addr->sun_path[sizeof(addr->sun_path) - 1], '\0');
-    }
+}
 
-    TEST(SocketDirectoryFromNamespace, Works) {
+TEST(SocketDirectoryFromNamespace, Works) {
   EXPECT_EQ(SocketDirectoryFromNamespace(""), "/tmp/intrinsic_icon");
 
   EXPECT_EQ(SocketDirectoryFromNamespace("some_namespace"),
@@ -90,29 +92,29 @@ namespace intrinsic::hal {
 
   EXPECT_EQ(SocketDirectoryFromNamespace("/some_namespace"),
             "/tmp/intrinsic_icon/some_namespace");
-    }
+}
 
-    TEST(GetSegmentNameToFileDescriptorMap, CreatesPath) {
+TEST(GetSegmentNameToFileDescriptorMap, CreatesPath) {
   auto socket_dir = std::filesystem::path(TempDir()) / "some_dir";
-      (void)GetSegmentNameToFileDescriptorMap(socket_dir, "some_module",
+  (void)GetSegmentNameToFileDescriptorMap(socket_dir, "some_module",
                                           std::chrono::seconds(0));
   std::error_code ec;
   EXPECT_TRUE(std::filesystem::exists(socket_dir, ec)) << ec;
   EXPECT_TRUE(std::filesystem::is_directory(socket_dir, ec)) << ec;
-    }
+}
 
-    TEST(GetSegmentNameToFileDescriptorMap, TimesOut) {
+TEST(GetSegmentNameToFileDescriptorMap, TimesOut) {
   auto result = GetSegmentNameToFileDescriptorMap(TempDir(), "some_module",
                                                   std::chrono::seconds(0));
   ASSERT_FALSE(result.has_value());
   EXPECT_EQ(result.error().code, StatusCode::kDeadlineExceeded);
-    }
+}
 
 // GetSegmentNameToFileDescriptorMap receiving descriptors and the version check
 // is validated in
 // intrinsic/icon/interprocess/shared_memory_manager/domain_socket_server_test.cc
 
-  } // namespace
+}   // namespace
 }  // namespace intrinsic::hal
 
 int main(int argc, char ** argv)
