+++ icon_hwm_controller/test/segment_header_test.cpp
@@ -10,36 +10,38 @@
 
 #include "intrinsic/utils/time.hpp"
 
-namespace intrinsic::hal {
+namespace intrinsic::hal
+{
 
-  using::testing::Eq;
-  using::testing::Ne;
-  using::testing::SizeIs;
-  using::testing::StrEq;
+using ::testing::Eq;
+using ::testing::Ne;
+using ::testing::SizeIs;
+using ::testing::StrEq;
 
 // Friend class to SegmentHeader, for testing purposes only. Refer to
 // go/totw/135 for more info on the "test peer" design pattern.
-  class SegmentHeaderTestPeer {
+class SegmentHeaderTestPeer {
 public:
-    static void SetUpdateCounter(int64_t value, SegmentHeader & header)
-    {
-      header.update_counter_ = value;
-    }
-    static void SetVersion(size_t version, SegmentHeader & header)
-    {
-      size_t * version_ptr = const_cast < size_t * > (&header.kVersion);
-      *version_ptr = version;
-    }
+  static void SetUpdateCounter(int64_t value, SegmentHeader & header)
+  {
+    header.update_counter_ = value;
+  }
+  static void SetVersion(size_t version, SegmentHeader & header)
+  {
+    size_t * version_ptr = const_cast<size_t *>(&header.kVersion);
+    *version_ptr = version;
+  }
 
-    static void SetUpdatedAtCycle(uint64_t value, SegmentHeader & header)
-    {
-      header.updated_at_cycle_ = value;
-    }
-  };
+  static void SetUpdatedAtCycle(uint64_t value, SegmentHeader & header)
+  {
+    header.updated_at_cycle_ = value;
+  }
+};
 
-  namespace {
+namespace
+{
 
-    TEST(SegmentHeaderTest, ModifyReferenceCounters) {
+TEST(SegmentHeaderTest, ModifyReferenceCounters) {
   SegmentHeader header;
   EXPECT_THAT(header.ReaderRefCount(), Eq(0));
   EXPECT_THAT(header.WriterRefCount(), Eq(0));
@@ -57,9 +59,9 @@ public:
   header.DecrementWriterRefCount();
   EXPECT_THAT(header.ReaderRefCount(), Eq(0));
   EXPECT_THAT(header.WriterRefCount(), Eq(0));
-    }
+}
 
-    TEST(SegmentHeaderTest, ReferenceCounterCantBeNegative) {
+TEST(SegmentHeaderTest, ReferenceCounterCantBeNegative) {
   SegmentHeader header;
   EXPECT_THAT(header.ReaderRefCount(), Eq(0));
   EXPECT_THAT(header.WriterRefCount(), Eq(0));
@@ -70,21 +72,21 @@ public:
   header.DecrementWriterRefCount();
   EXPECT_THAT(header.ReaderRefCount(), Eq(0));
   EXPECT_THAT(header.WriterRefCount(), Eq(0));
-    }
+}
 
-    TEST(SegmentHeaderTest, TypeInfoReturnsCorrectType) {
+TEST(SegmentHeaderTest, TypeInfoReturnsCorrectType) {
   SegmentHeader header("my_type");
   EXPECT_THAT(header.Type().TypeID(), StrEq("my_type"));
-    }
+}
 
-    TEST(SegmentHeaderTest, TypeInfoTruncatesTypeId) {
+TEST(SegmentHeaderTest, TypeInfoTruncatesTypeId) {
   const std::string kTooLongTypeId(SegmentHeader::TypeInfo::kMaxSize + 2, 'a');
   SegmentHeader header(kTooLongTypeId);
   EXPECT_THAT(header.Type().TypeID(),
               SizeIs(SegmentHeader::TypeInfo::kMaxSize));
-    }
+}
 
-    TEST(SegmentHeaderTest, TypeInfoComparesCorrectly) {
+TEST(SegmentHeaderTest, TypeInfoComparesCorrectly) {
   SegmentHeader header1("my_type");
   SegmentHeader header2("my_type");
 
@@ -92,9 +94,9 @@ public:
 
   SegmentHeader header3("my_other_type");
   EXPECT_THAT(header1.Type(), Ne(header3.Type()));
-    }
+}
 
-    TEST(SegmentHeaderTest, TruncatedTypeInfoComparesCorrectly) {
+TEST(SegmentHeaderTest, TruncatedTypeInfoComparesCorrectly) {
   const std::string kMaxSizeTypeId(SegmentHeader::TypeInfo::kMaxSize, 'a');
   SegmentHeader header1(kMaxSizeTypeId + "_first");
   SegmentHeader header2(kMaxSizeTypeId + "_second");
@@ -105,9 +107,9 @@ public:
   const std::string kOtherMaxSizeTypeId(SegmentHeader::TypeInfo::kMaxSize, 'b');
   SegmentHeader header3(kOtherMaxSizeTypeId);
   EXPECT_THAT(header1.Type(), Ne(header3.Type()));
-    }
+}
 
-    TEST(SegmentHeaderTest, QueryReturnsCorrectlySetFlags) {
+TEST(SegmentHeaderTest, QueryReturnsCorrectlySetFlags) {
   SegmentHeader no_flags("my_type1");
   EXPECT_FALSE(no_flags.FlagIsSet(SegmentHeader::Flags::kExclusiveOwnership));
 
@@ -115,13 +117,13 @@ public:
   EXPECT_FALSE(no_flags2.FlagIsSet(SegmentHeader::Flags::kExclusiveOwnership));
 
   SegmentHeader exclusive_flag(
-        "my_type",
-        {SegmentHeader::Flags::kExclusiveOwnership});
+    "my_type",
+    {SegmentHeader::Flags::kExclusiveOwnership});
   EXPECT_TRUE(
       exclusive_flag.FlagIsSet(SegmentHeader::Flags::kExclusiveOwnership));
-    }
+}
 
-    TEST(SegmentHeaderTest, UpdatedAt) {
+TEST(SegmentHeaderTest, UpdatedAt) {
   SegmentHeader my_header("my_type1");
   ASSERT_EQ(my_header.LastUpdatedTime(), ::intrinsic::Time());
   ASSERT_EQ(my_header.NumUpdates(), 0);
@@ -132,22 +134,22 @@ public:
   EXPECT_EQ(my_header.LastUpdatedTime(), now);
   EXPECT_EQ(my_header.NumUpdates(), 1);
   EXPECT_EQ(my_header.LastUpdatedCycle(), 42);
-    }
+}
 
-    TEST(SegmentHeaderTest, UpdatedAtOverrunWorks) {
+TEST(SegmentHeaderTest, UpdatedAtOverrunWorks) {
   SegmentHeader my_header("my_type");
 
-  int64_t initial_counter = std::numeric_limits < int64_t > ::max();
+  int64_t initial_counter = std::numeric_limits<int64_t>::max();
   SegmentHeaderTestPeer::SetUpdateCounter(initial_counter, my_header);
   EXPECT_EQ(my_header.NumUpdates(), initial_counter);
   // Overruns the counter without issues.
   // Validated using `-c dbg --config=ubsan`.
   my_header.UpdatedAt(::intrinsic::Now(), 0);
 
-  EXPECT_EQ(my_header.NumUpdates(), std::numeric_limits < int64_t > ::min());
-    }
+  EXPECT_EQ(my_header.NumUpdates(), std::numeric_limits<int64_t>::min());
+}
 
-    TEST(SegmentHeaderTest, VersionWorks) {
+TEST(SegmentHeaderTest, VersionWorks) {
   SegmentHeader my_header("my_type");
 
   EXPECT_EQ(my_header.Version(), SegmentHeader::ExpectedVersion());
@@ -156,9 +158,9 @@ public:
   SegmentHeaderTestPeer::SetVersion(SegmentHeader::ExpectedVersion() + 1,
                                     my_header);
   EXPECT_NE(my_header.Version(), SegmentHeader::ExpectedVersion());
-    }
+}
 
-  } // namespace
+}   // namespace
 }  // namespace intrinsic::hal
 
 int main(int argc, char ** argv)
